% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect_cube.R
\name{prospect_cube}
\alias{prospect_cube}
\title{Pull Prospect Rankings}
\usage{
prospect_cube(team, year, src = "BA", start_year = NULL, end_year = NULL)
}
\arguments{
\item{team}{Specify the team abbreviation as character value. Use this value to pull the top prospects by team}

\item{year}{Numeric year. Specifying the year will pull only the top 100 prospects. Default is to take the past full season if pulled in the offseason or the current season if pulled after opening day.}

\item{src}{Character value specifying the source to use for the prospect rankings. Options are `"BA"` for Baseball America or `"PIPE"` for MLB Pipeline. Default is Baseball America}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{end_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}
}
\description{
Scrape data from \url{http://www.thebaseballcube.com} to look at the top 100 prospects list for each team or the top prospects by team.
Can use this information to approximate the relative strength of each teams farm system.
}
\examples{
#pulling top 100 prospectes from 2017
top_prospects <- prospect_cube(year = 2017, src = "PIPE")
table(top_prospects$current_lvl)

#pulling prospect rankings from Colorado Rockies from Baseball America
col_prospects <- prospect_cube(team = "COL", src = "BA")
library(dplyr)
top_prospects <- col_prospects \%>\%
                   group_by(Year)\%>\%
                   summarise(num_top_100 = sum(MLB!=""), avg_age = mean(as.numeric(Age)))
top_prospects <- top_prospects[order(top_prospects$num_top_100, decreasing = TRUE),]
top_prospects

}
