% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect_fg.R
\name{prospect_fg}
\alias{prospect_fg}
\title{Pull FanGraphs Prospect Rankings from THE BOARD}
\usage{
prospect_fg(team, year, start_year = NULL, end_year = NULL, pos,
  scout_grades = FALSE)
}
\arguments{
\item{team}{Specify the team abbreviation as character value. Use this value to pull the top prospects by team}

\item{year}{Numeric year. Currently fangraphs has data only from 2017 and 2018.}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{end_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}

\item{pos}{Character value specifying a specific group of positions. By default position is set to all. See \url{https://www.fangraphs.com/scoutboard.aspx?draft=2018prospect&type=0&pos=all&team=all} for a list of all positions available.}

\item{scout_grades}{Indicator to pull the scout grades for each prospect. Based on the 20-80 scale. Default is not to pull the grades and only pull the value. Options are `TRUE` or `FALSE`.}
}
\description{
Scrape data from \url{https://www.fangraphs.com} to look at the top prospects based on ratings from FanGraphs prospect analysts
Can use this information to approximate the relative strength of each teams farm system.
}
\examples{
#pulling prospect rankings from the 2017 Colorado Rockies for all hitters
prospect_fg(year = 2017, team = "COL", pos = "BAT")

#pulling tool grades for 2017-2018 Los Angeles Dodgers
prospect_fg(start_year = 2017, end_year = 2018, team = "LAD", scout_grades = TRUE)

}
