% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_specific_fill.R
\name{team_specific_fill}
\alias{team_specific_fill}
\title{Finding league, division, and team specific extension to read data from Baseball-Reference}
\usage{
team_specific_fill(team)
}
\arguments{
\item{team}{Team abbreviation}
}
\value{
list that includes `league`, `division`, and `ext`
}
\description{
This function takes in a specific team abbreviation and returns the corresponding league, division, and their full name
}
