% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_contract.R
\name{tm_contract}
\alias{tm_contract}
\title{Pulling the contract information from Baseball-Reference}
\usage{
tm_contract(team)
}
\arguments{
\item{team}{Official team abbreviation}
}
\value{
Data.frame object that contains the salary information
}
\description{
Takes in the team abbreviation and returns their current and projected salary information based on 2017 data
}
\examples{
#reading in the Boston Red Sox salary
library(rwindow.baseball)
tm_contract("BOS")

}
