% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_player_stats.R
\name{tm_player_stats}
\alias{tm_player_stats}
\title{Pull the team stats for pitching and fielding}
\usage{
tm_player_stats(team, year = as.numeric(format(Sys.Date(), "\%Y")),
  start_year = NULL, end_year = NULL)
}
\arguments{
\item{team}{Team abbreviation}

\item{year}{Numeric year}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{start_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}
}
\description{
By default the system uses the current year
}
