% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plyr_batting.R
\name{tm_plyr_batting}
\alias{tm_plyr_batting}
\title{Scraping Fangraphs batter statistics}
\usage{
tm_plyr_batting(team, year, min_pa = 0, start_year = NULL,
  end_year = NULL)
}
\arguments{
\item{team}{Specify the team abbreviation}

\item{year}{Specify numeric year from which to pull}

\item{min_pa}{Minimum number of plate appearances for batters. Default is 0 and takes on values up to 1000. To see only qualified hitters use `y`.}

\item{start_year}{Numeric value indicating the first year from which you want to pull data}

\item{end_year}{Numeric value indicating the last year to pull for a year range.}
}
\description{
Allows the user to scrape batting statistics by team and season from fangraphs.com
}
