% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg_tm_value.R
\name{lg_tm_value}
\alias{lg_tm_value}
\title{Pull the value by team and season}
\usage{
lg_tm_value(league, year = as.numeric(format(Sys.Date(), "\%Y")),
  batting = T, start_year = NULL, end_year = NULL)
}
\arguments{
\item{league}{League abbreviation. Either AL or NL.}

\item{year}{Numeric year}

\item{batting}{Indicator variable to specify whether to pull batting or pitching value. By default it is set to batting.}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{start_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}
}
\description{
By default the system uses the current year
}
