% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect_fg.R
\name{prospect_fg}
\alias{prospect_fg}
\title{Pull FanGraphs Prospect Rankings from THE BOARD}
\usage{
prospect_fg(team, year, start_year = NULL, end_year = NULL, pos)
}
\arguments{
\item{team}{Specify the team abbreviation as character value. Use this value to pull the top prospects by team}

\item{year}{Numeric year. Currently fangraphs has data only from 2017 and 2018.}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{end_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}

\item{pos}{Character value specifying a specific group of positions. By default position is set to all. See \url{https://www.fangraphs.com/scoutboard.aspx?draft=2018prospect&type=0&pos=all&team=all} for a list of all positions available.}
}
\description{
Scrape data from \url{https://www.fangraphs.com} to look at the top prospects based on ratings from FanGraphs prospect analysts
Can use this information to approximate the relative strength of each teams farm system.
}
\examples{

}
