% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plyr_batting.R
\name{tm_plyr_batting}
\alias{tm_plyr_batting}
\title{Scraping Team Specific Fangraphs batter statistics}
\usage{
tm_plyr_batting(team, year, min_pa = 0, start_year = NULL,
  end_year = NULL)
}
\arguments{
\item{team}{Specify the team abbreviation as character value}

\item{year}{Specify numeric year from which to pull. Default is to take the current year if the season has started. Or the most recent year if the season is over.}

\item{min_pa}{Minimum number of plate appearances for batters. Default is 0 and takes on values up to 1000. To see only qualified hitters use `y`.}

\item{start_year}{Numeric value indicating the first year from which you want to pull data}

\item{end_year}{Numeric value indicating the last year to pull for a year range.}
}
\value{
Returns table with data.frame object on the team hitters with the main statistics referenced on the fangraphs leaderboard.
}
\description{
Allows the user to scrape leaderboard batting statistics by team and season from \url{http://www.fangraphs.com}.
}
\examples{
#reading in batters with at least 200 PA for the Boston Red Sox from 2003 to 2010
bos_batting <- tm_plyr_batting("BOS", start_year = 2003, end_year= 2010, min_pa = 200)
bos_batting[, c("AVG", "wOBA", "Season")] <- apply(bos_batting[, c("AVG", "wOBA", "Season")], 2, as.numeric)
head(bos_batting)

}
\seealso{
\code{\link{team_specific_fill}} for details on appropriate team values
}
