% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_standing_schedule.R
\name{tm_standings_schedule}
\alias{tm_standings_schedule}
\title{Pull the team standings and game by game schedule with results}
\usage{
tm_standings_schedule(team, year = as.numeric(format(Sys.Date(), "\%Y")),
  start_year = NULL, end_year = NULL)
}
\arguments{
\item{team}{Team abbreviation}

\item{year}{Numeric year}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{start_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}
}
\description{
By default the system uses the current year
}
