% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_standing_schedule.R
\name{tm_standings_schedule}
\alias{tm_standings_schedule}
\title{Scrape game results by team and season.}
\usage{
tm_standings_schedule(team, year, start_year = NULL, end_year = NULL)
}
\arguments{
\item{team}{Team abbreviation, three leter character value. See the data object \code{MLB_colors} for approved abbreviations.}

\item{year}{Numeric year. Default is to take the past full season if pulled in the offseason or the current season if pulled after opening day.}

\item{start_year}{Numeric value that identifies the beginning year to pull a range of data for the team of interest. This is an optional parameter.}

\item{end_year}{Numeric value that identifies the ending year to pull a range of data for the team of interest. This is an optional parameter.}
}
\value{

}
\description{
Pulls data on the team's opponent, result (W or L), runs allowed, runs scored, cumulative record, winning/losing pitcher last name,
time of game, type of game (day or night), attendance, and current win streak. Each row represents one game.
}
\details{
Data obtained from \url{https://www.baseball-reference.com}.
}
