% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_standing_schedule.R
\name{tm_standings_schedule}
\alias{tm_standings_schedule}
\title{Pull the team standings and game by game schedule with results}
\usage{
tm_standings_schedule(team, year = as.numeric(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{team}{Team abbreviation}

\item{year}{Numeric year}
}
\description{
By default the system uses the current year
}
